Add support for TFA and backup codes
It is possible to enable/disable TFA and fetch backup codes.
Login in for accounts with TFA enabled returns a token as well,
but the payload is diferent.
The payload of the token isn't checked yet when making a request,
which means that there is no need to authenticate with TFA yet.
However, that will change in the near future.